+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 
 ### Fixed
 - Service factory that used to randomly produce invalid objects (@tomek.bartynski)
+- Edit/destroy group buttons visible only for group owners (@mkasztelnik)
 
 ### Security